--- karma.conf.js
@@ -1,14 +1,0 @@
-module.exports = function(config) {
-    config.set({
-        frameworks: ['browserify','jasmine'],
-        files: ['src/**/*.js', 'test/**/*_spec.js'],
-        preprocessors: {
-            'test/**/*.js': ['jshint','browserify'],
-            'src/**/*.js': ['jshint','browserify']
-        },
-        browsers: ['PhantomJS'],
-        browserify: {
-			debug: true
-		}
-    })
-}
